// src/types/enums.ts
export enum CostElemType {
  Electric = 'Electric',
  Fire = 'Fire',
  Grass = 'Grass',
  Ice = 'Ice',
  None = 'None',
  Rock = 'Rock',
  Water = 'Water',
  Wind = 'Wind',
}
// ...
export enum FightProp {
  FightPropAddHurt = 'FIGHT_PROP_ADD_HURT',
  FightPropAntiCritical = 'FIGHT_PROP_ANTI_CRITICAL',
  FightPropAttack = 'FIGHT_PROP_ATTACK',
  FightPropAttackPercent = 'FIGHT_PROP_ATTACK_PERCENT',
  FightPropAttackPercentA = 'FIGHT_PROP_ATTACK_PERCENT_A',
  FightPropBaseAttack = 'FIGHT_PROP_BASE_ATTACK',
  FightPropBaseDefense = 'FIGHT_PROP_BASE_DEFENSE',
  FightPropBaseHP = 'FIGHT_PROP_BASE_HP',
  FightPropBaseSpeed = 'FIGHT_PROP_BASE_SPEED',
  FightPropChargeEfficiency = 'FIGHT_PROP_CHARGE_EFFICIENCY',
  FightPropCritical = 'FIGHT_PROP_CRITICAL',
  FightPropCriticalHurt = 'FIGHT_PROP_CRITICAL_HURT',
  FightPropCurAttack = 'FIGHT_PROP_CUR_ATTACK',
  FightPropCurDefense = 'FIGHT_PROP_CUR_DEFENSE',
  FightPropCurHP = 'FIGHT_PROP_CUR_HP',
  FightPropCurSpeed = 'FIGHT_PROP_CUR_SPEED',
  FightPropDefense = 'FIGHT_PROP_DEFENSE',
  FightPropDefensePercent = 'FIGHT_PROP_DEFENSE_PERCENT',
  FightPropDefensePercentA = 'FIGHT_PROP_DEFENSE_PERCENT_A',
  FightPropDizzyShorten = 'FIGHT_PROP_DIZZY_SHORTEN',
  FightPropEffectHit = 'FIGHT_PROP_EFFECT_HIT',
  FightPropEffectResist = 'FIGHT_PROP_EFFECT_RESIST',
  FightPropElecAddHurt = 'FIGHT_PROP_ELEC_ADD_HURT',
  FightPropElecSubHurt = 'FIGHT_PROP_ELEC_SUB_HURT',
  FightPropElementAddHurt = 'FIGHT_PROP_ELEMENT_ADD_HURT',
  FightPropElementMastery = 'FIGHT_PROP_ELEMENT_MASTERY',
  FightPropFireAddHurt = 'FIGHT_PROP_FIRE_ADD_HURT',
  FightPropFireSubHurt = 'FIGHT_PROP_FIRE_SUB_HURT',
  FightPropFreezeShorten = 'FIGHT_PROP_FREEZE_SHORTEN',
  FightPropGrassAddHurt = 'FIGHT_PROP_GRASS_ADD_HURT',
  FightPropGrassSubHurt = 'FIGHT_PROP_GRASS_SUB_HURT',
  FightPropHP = 'FIGHT_PROP_HP',
  FightPropHPPercent = 'FIGHT_PROP_HP_PERCENT',
  FightPropHPPercentA = 'FIGHT_PROP_HP_PERCENT_A',
  FightPropHealAdd = 'FIGHT_PROP_HEAL_ADD',
  FightPropHealedAdd = 'FIGHT_PROP_HEALED_ADD',
  FightPropIceAddHurt = 'FIGHT_PROP_ICE_ADD_HURT',
  FightPropIceSubHurt = 'FIGHT_PROP_ICE_SUB_HURT',
  FightPropMaxHP = 'FIGHT_PROP_MAX_HP',
  FightPropNone = 'FIGHT_PROP_NONE',
  FightPropPhysicalAddHurt = 'FIGHT_PROP_PHYSICAL_ADD_HURT',
  FightPropPhysicalSubHurt = 'FIGHT_PROP_PHYSICAL_SUB_HURT',
  FightPropRockAddHurt = 'FIGHT_PROP_ROCK_ADD_HURT',
  FightPropRockSubHurt = 'FIGHT_PROP_ROCK_SUB_HURT',
  FightPropShieldCostMinusRatio = 'FIGHT_PROP_SHIELD_COST_MINUS_RATIO',
  FightPropSkillCDMinusRatio = 'FIGHT_PROP_SKILL_CD_MINUS_RATIO',
  FightPropSpeedPercent = 'FIGHT_PROP_SPEED_PERCENT',
  FightPropSubHurt = 'FIGHT_PROP_SUB_HURT',
  FightPropTorporShorten = 'FIGHT_PROP_TORPOR_SHORTEN',
  FightPropWaterAddHurt = 'FIGHT_PROP_WATER_ADD_HURT',
  FightPropWaterSubHurt = 'FIGHT_PROP_WATER_SUB_HURT',
  FightPropWindAddHurt = 'FIGHT_PROP_WIND_ADD_HURT',
  FightPropWindSubHurt = 'FIGHT_PROP_WIND_SUB_HURT',
}
// ...
export enum BodyType {
  BodyBoy = 'BODY_BOY',
  BodyGirl = 'BODY_GIRL',
  BodyLady = 'BODY_LADY',
  BodyLoli = 'BODY_LOLI',
  BodyMale = 'BODY_MALE',
}
// ...
export enum WeaponType {
  WeaponBow = 'WEAPON_BOW',
  WeaponCatalyst = 'WEAPON_CATALYST',
  WeaponClaymore = 'WEAPON_CLAYMORE',
  WeaponPole = 'WEAPON_POLE',
  WeaponSwordOneHand = 'WEAPON_SWORD_ONE_HAND',
}